Adding WP CLI command

+++ lib/class-sp-sync-meta.php
@@ -1,7 +1,9 @@
 <?php
 
 /**
- * Simple class for working with the meta information associated with ES Syncing
+ * Simple class for working with the meta information associated with ES Syncing.
+ * Many of the methods are disabled if we're in the WP-CLI environment, but this
+ * class is still used to track the meta data (for consistency)
  *
  * @author Matthew Boynes
  */
@@ -40,10 +42,10 @@ class SP_Sync_Meta {
 
 
 	public function setup() {
-		if ( defined( 'SP_CLI' ) && SP_CLI )
-			return;
-
 		$this->init();
+
+		if ( defined( 'WP_CLI' ) && WP_CLI )
+			return;
 
 		if ( false != ( $sync_meta = get_option( 'sp_sync_meta' ) ) ) {
 			foreach ( $sync_meta as $key => $value ) {
@@ -74,7 +76,10 @@ class SP_Sync_Meta {
 	}
 
 
-	public function save() {
+	public function save( $force = false ) {
+		if ( ! 'force' == $force && defined( 'WP_CLI' ) && WP_CLI )
+			return;
+
 		$this->delete();
 		add_option( 'sp_sync_meta', array(
 			'running'   => $this->running,
@@ -90,7 +95,10 @@ class SP_Sync_Meta {
 	}
 
 
-	public function delete( $do_init = '' ) {
+	public function delete( $do_init = false, $force = false ) {
+		if ( ! 'force' == $force && defined( 'WP_CLI' ) && WP_CLI )
+			return;
+
 		delete_option( 'sp_sync_meta' );
 		if ( 'init' == $do_init )
 			$this->init();
@@ -98,6 +106,9 @@ class SP_Sync_Meta {
 
 
 	public function reload() {
+		if ( defined( 'WP_CLI' ) && WP_CLI )
+			return;
+
 		wp_cache_delete( 'sp_sync_meta', 'options' );
 		$this->setup();
 	}
